refactor: Refactored the name and root files of the project (CodeTrailer to CodeTour)
