+++ gradle.properties
@@ -1,8 +1,8 @@
 # IntelliJ Platform Artifacts Repositories
 # -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = org.uom.lefterisxris.trailer
-pluginName = CodeTrailer
+pluginGroup = org.uom.lefterisxris.codetour
+pluginName = CodeTour
 # SemVer format -> https://semver.org
 pluginVersion = 0.0.1
