+++ src/main/java/org/uom/lefterisxris/codetour/RunToursAction.java
@@ -1,14 +1,14 @@
-package org.uom.lefterisxris.trailer;
+package org.uom.lefterisxris.codetour;
 
 import com.intellij.openapi.actionSystem.ActionManager;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
-import org.uom.lefterisxris.trailer.tours.ToursDialog;
-import org.uom.lefterisxris.trailer.tours.RunTourByNameAction;
-import org.uom.lefterisxris.trailer.tours.ToursStateComponent;
-import org.uom.lefterisxris.trailer.tours.domain.Tour;
+import org.uom.lefterisxris.codetour.tours.ToursDialog;
+import org.uom.lefterisxris.codetour.tours.RunTourByNameAction;
+import org.uom.lefterisxris.codetour.tours.ToursStateComponent;
+import org.uom.lefterisxris.codetour.tours.domain.Tour;
 
 import java.util.ArrayList;
 import java.util.List;
