+++ src/main/java/org/uom/lefterisxris/codetour/SearchAction.java
@@ -1,4 +1,4 @@
-package org.uom.lefterisxris.trailer;
+package org.uom.lefterisxris.codetour;
 
 import com.intellij.ide.BrowserUtil;
 import com.intellij.lang.Language;
