+++ src/main/java/org/uom/lefterisxris/codetour/tours/ui/ToursNavigationWindow.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="org.uom.lefterisxris.trailer.tours.ui.ToursNavigationWindow">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="org.uom.lefterisxris.codetour.tours.ui.ToursNavigationWindow">
   <grid id="27dc6" binding="tourPanel" layout-manager="GridLayoutManager" row-count="2" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
