+++ src/main/kotlin/org/uom/lefterisxris/codetour/services/MyApplicationService.kt
@@ -1,6 +1,6 @@
-package org.uom.lefterisxris.trailer.services
+package org.uom.lefterisxris.codetour.services
 
-import org.uom.lefterisxris.trailer.MyBundle
+import org.uom.lefterisxris.codetour.MyBundle
 
 
 class MyApplicationService {
