+++ src/main/resources/META-INF/plugin.xml
@@ -1,67 +1,67 @@
 <!-- Plugin Configuration File. Read more: https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html -->
 <idea-plugin>
-   <id>org.uom.lefterisxris.trailer</id>
-   <name>CodeTrailer</name>
+   <id>org.uom.lefterisxris.codetour</id>
+   <name>CodeTour</name>
    <vendor>lefterisxris</vendor>
 
    <depends>com.intellij.modules.platform</depends>
 
    <extensions defaultExtensionNs="com.intellij">
-      <applicationService serviceImplementation="org.uom.lefterisxris.trailer.services.MyApplicationService"/>
-      <projectService serviceImplementation="org.uom.lefterisxris.trailer.services.MyProjectService"/>
-      <favoritesListProvider implementation="org.uom.lefterisxris.trailer.tours.TourFavoritesListProvider"/>
+      <applicationService serviceImplementation="org.uom.lefterisxris.codetour.services.MyApplicationService"/>
+      <projectService serviceImplementation="org.uom.lefterisxris.codetour.services.MyProjectService"/>
+      <favoritesListProvider implementation="org.uom.lefterisxris.codetour.tours.TourFavoritesListProvider"/>
       <toolWindow id="ToursNavigationWindow" anchor="left" icon="AllIcons.Actions.Colors"
-                  factoryClass="org.uom.lefterisxris.trailer.tours.ui.ToursNavigationWindowFactory"/>
+                  factoryClass="org.uom.lefterisxris.codetour.tours.ui.ToursNavigationWindowFactory"/>
    </extensions>
 
    <applicationListeners>
-      <listener class="org.uom.lefterisxris.trailer.listeners.MyProjectManagerListener"
+      <listener class="org.uom.lefterisxris.codetour.listeners.MyProjectManagerListener"
                 topic="com.intellij.openapi.project.ProjectManagerListener"/>
    </applicationListeners>
 
    <actions>
       <action
             id="StackOverflow.AskQuestion.ToolsMenu"
-            class="org.uom.lefterisxris.trailer.AskQuestionAction"
+            class="org.uom.lefterisxris.codetour.AskQuestionAction"
             text="Ask Question on Stack Overflow"
             description="Ask a Question on Stack Overflow">
          <add-to-group group-id="ToolsMenu" anchor="last"/>
       </action>
       <action
             id="StackOverflow.Search.Editor"
-            class="org.uom.lefterisxris.trailer.SearchAction"
+            class="org.uom.lefterisxris.codetour.SearchAction"
             text="Search on Stack Overflow"
             description="Search on Stack Overflow">
          <add-to-group group-id="EditorPopupMenu" anchor="last"/>
       </action>
       <action
             id="StackOverflow.Search.Console"
-            class="org.uom.lefterisxris.trailer.SearchAction"
+            class="org.uom.lefterisxris.codetour.SearchAction"
             text="Search on Stack Overflow"
             description="Search on Stack Overflow">
          <add-to-group group-id="ConsoleEditorPopupMenu" anchor="last"/>
       </action>
       <action id="RunToursAction"
-              class="org.uom.lefterisxris.trailer.RunToursAction"
+              class="org.uom.lefterisxris.codetour.RunToursAction"
               text="Run A Tour"
               description="Select a CodeTour to run, and let the IDE walk you through it">
          <add-to-group group-id="ToolsMenu" anchor="last"/>
       </action>
-      <action id="org.uom.lefterisxris.trailer.tours.TourGeneratorFromBreakpointsAction"
-              class="org.uom.lefterisxris.trailer.tours.TourGeneratorFromBreakpointsAction"
-              text="CodeTrailer - Generate Tours from BreakPoints"
+      <action id="org.uom.lefterisxris.codetour.tours.TourGeneratorFromBreakpointsAction"
+              class="org.uom.lefterisxris.codetour.tours.TourGeneratorFromBreakpointsAction"
+              text="CodeTour - Generate Tours from BreakPoints"
               description="Generate tours from breakPoints">
          <add-to-group group-id="ToolsMenu" anchor="first"/>
          <add-to-group group-id="Bookmarks" anchor="first"/>
       </action>
-      <action id="org.uom.lefterisxris.trailer.tours.TourStepGeneratorAction"
-              class="org.uom.lefterisxris.trailer.tours.TourStepGeneratorAction"
+      <action id="org.uom.lefterisxris.codetour.tours.TourStepGeneratorAction"
+              class="org.uom.lefterisxris.codetour.tours.TourStepGeneratorAction"
               text="Add Tour Step"
               description="Set this point a tour step">
          <add-to-group group-id="EditorGutterPopupMenu" anchor="first"/>
       </action>
-      <action id="org.uom.lefterisxris.trailer.PopupDialogAction"
-              class="org.uom.lefterisxris.trailer.PopupDialogAction"
+      <action id="org.uom.lefterisxris.codetour.PopupDialogAction"
+              class="org.uom.lefterisxris.codetour.PopupDialogAction"
               text="Action Basics Plugin: Pop Dialog Action"
               description="SDK action example"
               icon="AllIcons.Actions.Forward">
