+++ src/main/resources/messages/MyBundle.properties
@@ -1,3 +1,3 @@
-name=Code Trailer
+name=Code Tour
 applicationService=Application service
 projectService=Project service: {0}
